✅ extract the clock mocking test utility
